Some fix in ArticleParser
